+++ Chapter_06/scala/build.sbt
@@ -1,16 +1,25 @@
-name := "chapter04"
+name := "breeze.io"
 
 version := "1.0"
 
-scalaVersion := "2.10.4"
+libraryDependencies  ++= Seq(
+  "org.scalanlp" %% "breeze" % "0.12",
+  "org.scalanlp" %% "breeze-natives" % "0.12",
+  "org.scalanlp" %% "breeze-viz" % "0.12",
+  "org.apache.spark" %% "spark-core" % "1.6.1",
+  "org.apache.spark" % "spark-mllib_2.11" % "1.6.1",
+  "com.github.wookietreiber" %% "scala-chart" % "latest.integration",
+  "com.itextpdf" % "itextpdf" % "5.5.6",
+  "org.jfree" % "jfreesvg" % "3.0",
+  "com.databricks" % "spark-csv_2.11" % "1.4.0"
+)
 
-libraryDependencies += "org.apache.spark" %% "spark-core" % "1.6.0"
-libraryDependencies += "org.scalanlp" %% "breeze" % "0.12"
-// native libraries are not included by default. add this if you want them (as of 0.7)
-// native libraries greatly improve performance, but increase jar sizes.
-// It also packages various blas implementations, which have licenses that may or may not
-// be compatible with the Apache License. No GPL code, as best I know.
-libraryDependencies +="org.scalanlp" %% "breeze-natives" % "0.12"
-libraryDependencies +="org.jfree" % "jfreechart" % "1.0.14"
-libraryDependencies += "com.github.wookietreiber" %% "scala-chart" % "0.3.0"
-libraryDependencies += "org.apache.spark" % "spark-mllib_2.10" % "1.6.0"
+
+resolvers ++= Seq(
+  // other resolvers here
+  // if you want to use snapshot builds (currently 0.12-SNAPSHOT), use this.
+  "Sonatype Snapshots" at "https://oss.sonatype.org/content/repositories/snapshots/",
+  "Sonatype Releases" at "https://oss.sonatype.org/content/repositories/releases/"
+)
+
+scalaVersion := "2.11.7"
