--- Chapter_06/scala/src/main/scala/org/sparksamples/classification/GradientBoostedTreesImpl.scala
@@ -1,49 +1,0 @@
-package org.sparksamples.classification
-
-import org.apache.spark.SparkContext
-import org.apache.spark.mllib.tree.GradientBoostedTrees
-import org.apache.spark.mllib.tree.configuration.BoostingStrategy
-import org.apache.spark.mllib.tree.model.GradientBoostedTreesModel
-import org.apache.spark.mllib.util.MLUtils
-
-/**
-  * Created by manpreet.singh on 14/04/16.
-  */
-object GradientBoostedTreesImpl {
-
-  def main(args: Array[String]) {
-    val sc = new SparkContext("local[2]", "Classification")
-
-    // Load and parse the data file.
-    val data = MLUtils.loadLibSVMFile(sc, "/Users/manpreet.singh/Sandbox/codehub/github/machinelearning/breeze.io/src/main/scala/moviestream/classification/sample_libsvm_data.txt")
-    // Split the data into training and test sets (30% held out for testing)
-    val splits = data.randomSplit(Array(0.7, 0.3))
-    val (trainingData, testData) = (splits(0), splits(1))
-
-    // Train a GradientBoostedTrees model.
-    // The defaultParams for Classification use LogLoss by default.
-    val boostingStrategy = BoostingStrategy.defaultParams("Classification")
-    boostingStrategy.numIterations = 3 // Note: Use more iterations in practice.
-    boostingStrategy.treeStrategy.numClasses = 2
-    boostingStrategy.treeStrategy.maxDepth = 5
-    // Empty categoricalFeaturesInfo indicates all features are continuous.
-    boostingStrategy.treeStrategy.categoricalFeaturesInfo = Map[Int, Int]()
-
-    val model = GradientBoostedTrees.train(trainingData, boostingStrategy)
-
-    // Evaluate model on test instances and compute test error
-    val labelAndPreds = testData.map { point =>
-      val prediction = model.predict(point.features)
-      (point.label, prediction)
-    }
-    val testErr = labelAndPreds.filter(r => r._1 != r._2).count.toDouble / testData.count()
-    println("Test Error = " + testErr)
-    println("Learned classification GBT model:\n" + model.toDebugString)
-
-    // Save and load model
-    model.save(sc, "target/tmp/myGradientBoostingClassificationModel")
-    val sameModel = GradientBoostedTreesModel.load(sc,
-      "target/tmp/myGradientBoostingClassificationModel")
-  }
-
-}
