--- Chapter_06/scala/src/main/scala/org/sparksamples/classification/RandomForest.scala
@@ -1,49 +1,0 @@
-package org.sparksamples.classification
-
-import org.apache.spark.SparkContext
-import org.apache.spark.mllib.tree.RandomForest
-import org.apache.spark.mllib.tree.model.RandomForestModel
-import org.apache.spark.mllib.util.MLUtils
-
-/**
-  * Created by manpreet.singh on 14/04/16.
-  */
-object RandomForestImpl {
-
-  def main(args: Array[String]) {
-    val sc = new SparkContext("local[2]", "Classification")
-
-    // Load and parse the data file.
-    val data = MLUtils.loadLibSVMFile(sc, "/Users/manpreet.singh/Sandbox/codehub/github/machinelearning/breeze.io/src/main/scala/moviestream/classification/sample_libsvm_data.txt")
-    // Split the data into training and test sets (30% held out for testing)
-    val splits = data.randomSplit(Array(0.7, 0.3))
-    val (trainingData, testData) = (splits(0), splits(1))
-
-    // Train a RandomForest model.
-    // Empty categoricalFeaturesInfo indicates all features are continuous.
-    val numClasses = 2
-    val categoricalFeaturesInfo = Map[Int, Int]()
-    val numTrees = 3 // Use more in practice.
-    val featureSubsetStrategy = "auto" // Let the algorithm choose.
-    val impurity = "gini"
-    val maxDepth = 4
-    val maxBins = 32
-
-    val model = RandomForest.trainClassifier(trainingData, numClasses, categoricalFeaturesInfo,
-      numTrees, featureSubsetStrategy, impurity, maxDepth, maxBins)
-
-    // Evaluate model on test instances and compute test error
-    val labelAndPreds = testData.map { point =>
-      val prediction = model.predict(point.features)
-      (point.label, prediction)
-    }
-    val testErr = labelAndPreds.filter(r => r._1 != r._2).count.toDouble / testData.count()
-    println("Test Error = " + testErr)
-    println("Learned classification forest model:\n" + model.toDebugString)
-
-    // Save and load model
-    model.save(sc, "target/tmp/myRandomForestClassificationModel")
-    val sameModel = RandomForestModel.load(sc, "target/tmp/myRandomForestClassificationModel")
-  }
-
-}
