+++ Chapter_06/scala/src/main/scala/org/sparksamples/classification/stumbleupon/SparkCommonUtils.scala
@@ -1,0 +1,14 @@
+package org.sparksamples.classification.stumbleupon
+
+import org.apache.spark.SparkConf
+
+/**
+  * Created by manpreet.singh on 26/04/16.
+  */
+object SparkCommonUtils {
+
+  def createSparkConf(appName: String): SparkConf = {
+     new SparkConf().setAppName(appName).setMaster(SparkConstants.SparkMaster)
+  }
+
+}
