+++ Chapter_06/scala/src/main/scala/org/sparksamples/classification/stumbleupon/SparkConstants.scala
@@ -1,0 +1,10 @@
+package org.sparksamples.classification.stumbleupon
+
+/**
+  * Created by manpreet.singh on 26/04/16.
+  */
+object SparkConstants {
+
+  val SparkMaster = "local[2]"
+
+}
